+++ src/stateMachine/query/queryMachine.ts
@@ -1,14 +1,14 @@
 import type { TypeDBDriver } from 'typedb-driver';
-import type { SurrealClient } from '../../adapters/surrealDB/client';
+import type { SurrealClient } from '../../adapters/surrealdb/client';
+import { runSurrealDbQueryMachine2 } from '../../adapters/surrealdb/query/run';
+import { runTypeDbQueryMachine } from '../../adapters/typedb/query/machine';
+import { enrichBQLQuery } from '../../bql/query/enrich';
 import { getSchemaByThing } from '../../helpers';
 import { logDebug } from '../../logger';
 import { createMachine, interpret, invoke, reduce, state, transition } from '../../robot3';
 import type { BormConfig, DBHandles, EnrichedBormSchema, RawBQLQuery } from '../../types';
 import type { DRAFT_EnrichedBormSchema } from '../../types/schema/enriched.draft';
 import { VERSION } from '../../version';
-import { enrichBQLQuery } from './bql/enrich';
-import { runSurrealDbQueryMachine2 } from './surql2/run';
-import { runTypeDbQueryMachine } from './tql/machine';
 
 type MachineContext = {
   bql: {